callbackを切り出してみる
